Refactor move size class declarations to androidx package
