+++ app/src/desktopMain/kotlin/io/github/xxfast/androidx/compose/material3/windowsizeclass/CalculateWindowSizeClass.kt
@@ -1,8 +1,8 @@
-package io.github.xxfast.nytimes.utils
+package io.github.xxfast.androidx.compose.material3.windowsizeclass
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.window.WindowState
-import io.github.xxfast.nytimes.utils.CommonWindowSizeClass.Companion.calculateFromSize
+import io.github.xxfast.androidx.compose.material3.windowsizeclass.CommonWindowSizeClass.Companion.calculateFromSize
 
 @Composable
 fun calculateWindowSizeClass(windowState: WindowState): WindowSizeClass = calculateFromSize(windowState.size)
