+++ app/src/desktopMain/kotlin/io/github/xxfast/androidx/compose/material3/windowsizeclass/WindowHeightSizeClass.kt
@@ -1,8 +1,6 @@
-package io.github.xxfast.nytimes.utils
+package io.github.xxfast.androidx.compose.material3.windowsizeclass
 
-import androidx.compose.material3.windowsizeclass.WindowHeightSizeClass
-
-actual typealias WindowHeightSizeClass = WindowHeightSizeClass
+actual typealias WindowHeightSizeClass = CommonWindowHeightSizeClass
 
 actual object WindowHeightSizeClasses {
   actual val Compact: WindowHeightSizeClass = WindowHeightSizeClass.Compact
