+++ app/src/desktopMain/kotlin/io/github/xxfast/androidx/compose/material3/windowsizeclass/WindowSizeClass.kt
@@ -1,0 +1,7 @@
+package io.github.xxfast.androidx.compose.material3.windowsizeclass
+
+actual typealias WindowSizeClass = CommonWindowSizeClass
+
+
+
+
