+++ app/src/iosMain/kotlin/io/github/xxfast/androidx/compose/material3/windowsizeclass/WindowSizeClass.kt
@@ -1,0 +1,3 @@
+package io.github.xxfast.androidx.compose.material3.windowsizeclass
+
+actual typealias WindowSizeClass = CommonWindowSizeClass
