--- app/src/iosMain/kotlin/io/github/xxfast/nytimes/utils/WindowSizeClass.kt
@@ -1,3 +1,0 @@
-package io.github.xxfast.nytimes.utils
-
-actual typealias WindowSizeClass = CommonWindowSizeClass
