+++ app/src/jsMain/kotlin/io/github/xxfast/androidx/compose/material3/windowsizeclass/WindowHeightSizeClass.kt
@@ -1,4 +1,4 @@
-package io.github.xxfast.nytimes.utils
+package io.github.xxfast.androidx.compose.material3.windowsizeclass
 
 actual typealias WindowHeightSizeClass = CommonWindowHeightSizeClass
