+++ app/src/jsMain/kotlin/io/github/xxfast/androidx/compose/material3/windowsizeclass/WindowSizeClass.kt
@@ -1,0 +1,9 @@
+package io.github.xxfast.androidx.compose.material3.windowsizeclass
+
+import io.github.xxfast.androidx.compose.material3.windowsizeclass.CommonWindowSizeClass
+
+actual typealias WindowSizeClass = CommonWindowSizeClass
+
+
+
+
