--- app/src/jsMain/kotlin/io/github/xxfast/nytimes/utils/CalculateWindowSizeClass.kt
@@ -1,8 +1,0 @@
-package io.github.xxfast.nytimes.utils
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.unit.DpSize
-
-@Composable
-fun calculateWindowSizeClass(size: DpSize): WindowSizeClass = WindowSizeClass.calculateFromSize(size)
-
