+++ app/web/src/jsMain/kotlin/io/github/xxfast/nytimes/web/Application.kt
@@ -8,9 +8,9 @@ import com.arkivanov.decompose.DefaultComponentContext
 import com.arkivanov.essenty.lifecycle.LifecycleRegistry
 import io.github.xxfast.decompose.LocalComponentContext
 import io.github.xxfast.nytimes.screens.home.HomeScreen
-import io.github.xxfast.nytimes.utils.LocalWindowSizeClass
-import io.github.xxfast.nytimes.utils.WindowSizeClass
-import io.github.xxfast.nytimes.utils.calculateWindowSizeClass
+import io.github.xxfast.androidx.compose.material3.windowsizeclass.LocalWindowSizeClass
+import io.github.xxfast.androidx.compose.material3.windowsizeclass.WindowSizeClass
+import io.github.xxfast.androidx.compose.material3.windowsizeclass.calculateWindowSizeClass
 import io.github.xxfast.nytimes.web.utils.BrowserViewportWindow
 import org.jetbrains.skiko.wasm.onWasmReady
